Rename AddDayofWeekNearTaskBarClock to DayClock.ps1
This file can be run standalone to have the day of week display near the windows clock in lower right hand corner. It was made as someone wanted this feature in the taskbar however there was no easy windows API I was aware of that could alter the taskbar functions.